new stream textures.
Former-commit-id: 7c6ff509f0028d94ca0d7d550ff9191edcfa7f8b [formerly ba2525d751ff0dd067ef3d6ead495c5d3107b639]
Former-commit-id: 69a90a70603ed8e0b29e8bf7d4b5ff2bf92cdc20

+++ Source/Editor/editor.h
@@ -58,7 +58,7 @@ private:
 
 	void initializeWindow();
 	void initializeControls();
-	void handleScroll(int a_amount);
+	void handleScroll(float a_amount);
 
 	Managers& managers;
